feat: only allow lowercase slug configuration (#11)
* feat: check for slugs in a case-insensitive way

"example-1" and "EXAMPLE-1" should be counted as duplicated. See feedback: https://github.com/privacybydesign/yivi-generic-identity-attestation-provider/pull/6#discussion_r2161546833

* test: add case-insensitive duplicate slug test

Also update the original test with the new exception message.

* refactor: only allow lowercase slugs

For user input convert the slugs to lowercase.
This is necessary because when a user navigates to /Example, it shouldn't respond that it didn't find anything.

* docs: remove implementation detail comment from summary docs